+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - [9955](https://github.com/vegaprotocol/vega/issues/9955) - Add data node subscription for transaction results.
 - [10004](https://github.com/vegaprotocol/vega/issues/10004) Track average entry price in position engine
 - [9825](https://github.com/vegaprotocol/vega/issues/9825) - Remove quadratic slippage.
+- [9943](https://github.com/vegaprotocol/vega/issues/9943) - Support amending the order size by defining the target size.
 
 ### 🐛 Fixes